+++ LearnOpenGL/src/VertexBuffer.cpp
@@ -1,6 +1,6 @@
 #include "VertexBuffer.h"
 
-VertexBuffer::VertexBuffer(const void* data, unsigned int size, GLenum mode)
+VertexBuffer::VertexBuffer(const void* data, size_t size, GLenum mode)
 {
     glGenBuffers(1, &id);
     glBindBuffer(GL_ARRAY_BUFFER, id);
